# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Deni",
        "description": "Descripción de la API de Ejemplo",
        "termsOfService": "https://www.ejemplo.com/terms",
        "contact": {
            "email": "contacto@ejemplo.com"
        },
        "license": {
            "name": "Licencia de Ejemplo",
            "url": "https://www.ejemplo.com/licencia"
        },
        "version": "1.0"
    },
    "paths": {
        "/adoptantes": {
            "get": {
                "tags": [
                    "Adoptantes"
                ],
                "summary": "Obtener todos los adoptantes",
                "description": "Recupera todos los adoptantes de la base de datos.",
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "Lista de adoptantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adoptante"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Adoptantes"
                ],
                "summary": "Registrar un nuevo adoptante",
                "description": "Crear un nuevo adoptante con la información proporcionada.",
                "operationId": "store",
                "requestBody": {
                    "description": "Datos del adoptante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "correo_electronico",
                                    "telefono",
                                    "direccion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nombre"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Apellido"
                                    },
                                    "correo_electronico": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "adoptante@ejemplo.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456789"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Dirección de ejemplo"
                                    },
                                    "ocupacion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Ocupación de ejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Adoptante registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adoptante"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/adoptantes/{id}": {
            "get": {
                "tags": [
                    "Adoptantes"
                ],
                "summary": "Obtener información de un adoptante",
                "description": "Obtiene información de un adoptante por su ID.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del adoptante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del adoptante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adoptante"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Adoptante no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Adoptante no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animales": {
            "get": {
                "tags": [
                    "Animales"
                ],
                "summary": "Obtener todos los animales",
                "description": "Recupera todos los animales de la base de datos.",
                "operationId": "indexAnimales",
                "responses": {
                    "200": {
                        "description": "Lista de animales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/animal"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Animales"
                ],
                "summary": "Registrar un nuevo animal",
                "description": "Crear un nuevo animal con la información proporcionada.",
                "operationId": "storeAnimal",
                "requestBody": {
                    "description": "Datos del animal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "especie"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nombre"
                                    },
                                    "especie": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Especie"
                                    },
                                    "raza": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Raza"
                                    },
                                    "edad": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": 3
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Descripción del animal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Animal registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/animal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/animales/{id}": {
            "get": {
                "tags": [
                    "Animales"
                ],
                "summary": "Obtener información de un animal",
                "description": "Obtiene información de un animal por su ID.",
                "operationId": "showAnimal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del animal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del animal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/animal"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Animal no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Animal no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Iniciar sesión en la aplicación.",
                "operationId": "login",
                "requestBody": {
                    "description": "Credenciales de inicio de sesión",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_usuario",
                                    "contrasena"
                                ],
                                "properties": {
                                    "nombre_usuario": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "usuarioejemplo"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasenaejemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "format": "int"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nombre_usuario": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "contrasena": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Registro"
                ],
                "summary": "Registrarse",
                "description": "Registrar una nueva cuenta de usuario.",
                "operationId": "register",
                "requestBody": {
                    "description": "Datos de registro de usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "nombre_usuario",
                                    "contrasena",
                                    "contrasena_confirmation",
                                    "apellido",
                                    "correo_electronico",
                                    "rol"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Nombre"
                                    },
                                    "nombre_usuario": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "usuarioejemplo"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasenaejemplo"
                                    },
                                    "contrasena_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "contrasenaejemplo"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Apellido"
                                    },
                                    "correo_electronico": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@ejemplo.com"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Rol"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "usuario": {
                                            "$ref": "#/components/schemas/usuario"
                                        },
                                        "token": {
                                            "type": "string",
                                            "format": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en el registro"
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nombre": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "nombre_usuario": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "contrasena": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "apellido": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "correo_electronico": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "rol": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/getaccount": {
            "get": {
                "tags": [
                    "usuario"
                ],
                "summary": "Obtener información del usuario autenticado",
                "description": "Obtiene información del usuario que ha iniciado sesión.",
                "operationId": "getaccount",
                "responses": {
                    "200": {
                        "description": "Información del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/usuario"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Cerrar sesión del usuario autenticado.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Sesión cerrada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener todos los usuarios",
                "description": "Obtiene la lista de todos los usuarios en la base de datos",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/usuario"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/generarPdf": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Generar PDF de usuarios",
                "description": "Genera un archivo PDF que contiene la lista de usuarios",
                "operationId": "generarPDF",
                "responses": {
                    "200": {
                        "description": "PDF generado exitosamente"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "adoptante": {
                "title": "adoptante",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "correo_electronico": {
                        "type": "string"
                    },
                    "telefono": {
                        "type": "string"
                    },
                    "direccion": {
                        "type": "string"
                    },
                    "ocupacion": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "animal": {
                "title": "animal",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "especie": {
                        "type": "string"
                    },
                    "raza": {
                        "type": "string"
                    },
                    "edad": {
                        "type": "integer"
                    },
                    "descripcion": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "usuario": {
                "title": "usuario",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "nombre_usuario": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    },
                    "correo_electronico": {
                        "type": "string"
                    },
                    "rol": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}